Remove use of controllers to index prefixes
This is effectivly a revert of PR #642
the reasoning is this caused subtle bugs with
model changes that did not go via a controller.

The problem this caused was that the database would be out of sync
with the index.

This was most notable in repository transfers where the DB
would show correct transfers but index remained unchanged.

To compensate for the original bugfix, we just now index directly rather than wait
for an indexjob