Fix to lookup correct object_key when new object version has not been created.
When the metadata version was set on before_validation it made a new version that was going to be stored before the data is actually stored. This meant when object_key was looking up a new version when it didn't exist yet.
This might occur when testing validation which happens before create lifecycle callback correctly stores the new version of the metadata.

In addition also added some nil checks to prevent lookups of data where it didn't exist thus also failing validation, e.g. when building an object and not actually storing it.